cannon: Use uint64 over int

+++ cannon/mipsevm/debug.go
@@ -8,11 +8,11 @@ type DebugInfo struct {
 	NumPreimageRequests int            `json:"num_preimage_requests"`
 	TotalPreimageSize   int            `json:"total_preimage_size"`
 	//  Multithreading-related stats below
-	RmwSuccessCount              int            `json:"rmw_success_count"`
-	RmwFailCount                 int            `json:"rmw_fail_count"`
+	RmwSuccessCount              uint64         `json:"rmw_success_count"`
+	RmwFailCount                 uint64         `json:"rmw_fail_count"`
 	MaxStepsBetweenLLAndSC       hexutil.Uint64 `json:"max_steps_between_ll_and_sc"`
-	ReservationInvalidationCount int            `json:"reservation_invalidation_count"`
-	ForcedPreemptionCount        int            `json:"forced_preemption_count"`
-	FailedWakeupCount            int            `json:"failed_wakeup_count"`
+	ReservationInvalidationCount uint64         `json:"reservation_invalidation_count"`
+	ForcedPreemptionCount        uint64         `json:"forced_preemption_count"`
+	FailedWakeupCount            uint64         `json:"failed_wakeup_count"`
 	IdleStepCountThread0         hexutil.Uint64 `json:"idle_step_count_thread_0"`
 }
